refactor: embed gorm.Model in domain entities for timestamp standardization
- Embed gorm.Model in review, shipping, market, product, coupon domain entities
- Update repositories with toEntity/fromEntity conversions for embedded gorm.Model
- Update services and DTOs for uint ID and time.Time timestamp types
- Add upload.api import to admin.api and fix multipart.FileHeader type assertion
- Update db.md rules to mandate gorm.Model embedding pattern

+++ admin/internal/application/review/dto.go
@@ -105,7 +105,7 @@ type BatchOperationResult struct {
 // FromDomainReview converts domain review to DTO
 func FromDomainReview(r *review.Review) *ReviewDetailDTO {
 	dto := &ReviewDetailDTO{
-		ID:            r.ID,
+		ID:            int64(r.ID),
 		TenantID:      r.TenantID.Int64(),
 		OrderID:       r.OrderID,
 		ProductID:     r.ProductID,
@@ -128,7 +128,7 @@ func FromDomainReview(r *review.Review) *ReviewDetailDTO {
 
 	if r.Reply != nil {
 		dto.Reply = &ReplyDTO{
-			ID:        r.Reply.ID,
+			ID:        int64(r.Reply.ID),
 			Content:   r.Reply.Content,
 			AdminName: r.Reply.AdminName,
 			CreatedAt: r.Reply.CreatedAt.Format(time.RFC3339),
