+++ admin/internal/logic/shipping_templates/update_shipping_template_logic.go
@@ -59,7 +59,7 @@ func (l *UpdateShippingTemplateLogic) UpdateShippingTemplate(req *types.UpdateSh
 	}
 
 	return &types.ShippingTemplateDetailResp{
-		ID:        template.ID,
+		ID:        int64(template.ID),
 		Name:      template.Name,
 		IsDefault: template.IsDefault,
 		IsActive:  template.IsActive,
