+++ admin/internal/logic/shipping_zones/update_shipping_zone_logic.go
@@ -87,7 +87,7 @@ func (l *UpdateShippingZoneLogic) UpdateShippingZone(req *types.UpdateShippingZo
 	}
 
 	return &types.ShippingZoneDetail{
-		ID:                  zone.ID,
+		ID:                  int64(zone.ID),
 		TemplateID:          zone.TemplateID,
 		Name:                zone.Name,
 		Regions:             zone.Regions,
